requirements.txt
quart==0.20.0
hypercorn==0.17.3
websockets==16.0
python-dotenv==1.2.2
requests==2.32.3
pillow==11.1.0
